Move InfoMethods from statistics to info directory
