+++ OOVisualization/OOVisualization.pro
@@ -120,7 +120,7 @@ HEADERS += src/statements/VAssertStatementStyle.h \
     src/statements/VSynchronizedStatementStyle.h \
     src/expressions/VCastExpression.h \
     src/expressions/VCastExpressionStyle.h \
-    src/statistics/InfoMethods.h
+    src/info/InfoMethods.h
 SOURCES += src/statements/VAssertStatementStyle.cpp \
     src/statements/VAssertStatement.cpp \
     src/expressions/VSuperExpression.cpp \
@@ -230,4 +230,4 @@ SOURCES += src/statements/VAssertStatementStyle.cpp \
     src/statements/VSynchronizedStatementStyle.cpp \
     src/expressions/VCastExpression.cpp \
     src/expressions/VCastExpressionStyle.cpp \
-    src/statistics/InfoMethods.cpp
+    src/info/InfoMethods.cpp
