+++ OOVisualization/src/OOVisualizationPlugin.cpp
@@ -37,7 +37,7 @@
 #include "VisualizationBase/src/node_extensions/Position.h"
 #include "VisualizationBase/src/node_extensions/FullDetailSize.h"
 
-#include "statistics/InfoMethods.h"
+#include "info/InfoMethods.h"
 #include "VisualizationBase/src/nodes/InfoNode.h"
 
 using namespace OOModel;
@@ -126,8 +126,8 @@ bool OOVisualizationPlugin::initialize(Core::EnvisionManager&)
 	Scene::defaultRenderer()->registerVisualization(Project::typeIdStatic(), "default_purpose",
 		"test", createVisualization<VDeclarationConstantSz, Project>, VDeclarationConstantSz::typeIdStatic());
 
-	Visualization::InfoNode::registerInfoGetter("callees", InfoMethods::numberOfCallees, false, false);
-	Visualization::InfoNode::registerInfoGetter("name", InfoMethods::fullName, true, true);
+	Visualization::InfoNode::registerInfoGetter("callees", InfoMethods::numberOfCallees, true, true);
+	Visualization::InfoNode::registerInfoGetter("name", InfoMethods::fullName, false, true);
 
 	return true;
 }
