+++ OOVisualization/src/info/InfoMethods.cpp
